requirements.txt
acitoolkit==0.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
deepdiff==3.3.0
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-SQLAlchemy==2.3.0
Flask-WTF==0.14.2
functools32==3.2.3.post2
gitdb2==2.0.3
GitPython==2.1.7
graphviz==0.8
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
jsonpickle==0.9.5
jsonschema==2.6.0
MarkupSafe==1.0
py-radix==0.9.7
PyMySQL==0.7.11
requests==2.18.4
six==1.11.0
smmap2==2.0.3
SQLAlchemy==1.1.14
tabulate==0.8.0
urllib3==1.22
visitor==0.1.3
websocket-client==0.44.0
Werkzeug==0.12.2
WTForms==2.1
